systems/mapeo/cordoba-dataset/requirements.txt
pandas
requests
pyyaml
python-dotenv